requirements.txt
streamlit
pandas
pdfplumber
pytesseract
docx2txt
PyMuPDF
sentence-transformers
Pillow
spacy
torch  # Required by sentence-transformers
scikit-learn  # Required by sentence-transformers